requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg[binary]==3.2.3
django-auth-ldap==5.1.0
python-ldap==3.4.4
holidays>=0.56,<1.0
